[dg] Fold `dg init` into `dg scaffold project`
