+++ docs/docs/guides/labs/dg/multiple-projects.md
@@ -2,7 +2,7 @@
 title: 'Managing multiple projects with dg'
 sidebar_label: 'Managing multiple projects'
 sidebar_position: 300
-description: Manage multiple isolated Dagster projects using dg, each with unique environments, by creating a workspace directory with dg init.
+description: Manage multiple isolated Dagster projects using dg, each with unique environments, by creating a workspace directory with dg scaffold project.
 ---
 
 import DgComponentsPreview from '@site/docs/partials/\_DgComponentsPreview.md';
@@ -27,9 +27,14 @@ A workspace does not define a Python environment by default. Instead, Python env
 
 ## Scaffold a new workspace and first project
 
-To scaffold a new workspace with an initial project called `project-1`, run `dg init` with the `--workspace` option and `--python-environment uv_managed` option. You will be prompted for the name of the project:
+To scaffold a new workspace called `dagster-workspace`, run `dg scaffold
+workspace`:
 
-<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/1-dg-init.txt" />
+<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/1-dg-scaffold-workspace.txt" />
+
+Now we'll create a project inside our workspace called `project-1`. Run `dg scaffold project` with the `--python-environment uv_managed` option. You will be prompted for the name of the project:
+
+<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/2-dg-scaffold-project.txt" />
 
 :::note
 
@@ -43,12 +48,12 @@ This will create a new directory called `dagster-workspace` with a `projects` su
 
 The new workspace has the following structure:
 
-<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/2-tree.txt" />
+<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/3-tree.txt" />
 
 The `dg.toml` file for the `dagster-workspace` folder contains a `directory_type = "workspace"` setting that marks this directory as a workspace:
 
 <CodeExample
-  path="docs_snippets/docs_snippets/guides/dg/workspace/3-dg.toml"
+  path="docs_snippets/docs_snippets/guides/dg/workspace/4-dg.toml"
   language="TOML"
   title="dagster-workspace/dg.toml"
 />
@@ -63,7 +68,7 @@ The `project-1` directory contains a `pyproject.toml` file with a
 `tool.dg.directory_type = "project"` section that defines it as a `dg` project:
 
 <CodeExample
-  path="docs_snippets/docs_snippets/guides/dg/workspace/4-project-pyproject.toml"
+  path="docs_snippets/docs_snippets/guides/dg/workspace/5-project-pyproject.toml"
   language="TOML"
   title="dagster-workspace/projects/project-1/pyproject.toml"
 />
@@ -74,11 +79,11 @@ As noted above, environments are scoped per project. `dg` commands will only use
 
 Let's create another project:
 
-<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/5-scaffold-project.txt" />
+<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/6-scaffold-project.txt" />
 
 Now we have two projects. We can list them with:
 
-<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/6-project-list.txt" />
+<CliInvocationExample path="docs_snippets/docs_snippets/guides/dg/workspace/7-project-list.txt" />
 
 ## Load workspace with `dg`
