+++ examples/docs_snippets/docs_snippets/guides/components/index/2-a-uv-scaffold.txt
@@ -1,1 +1,1 @@
-dg init jaffle-platform
+dg scaffold project jaffle-platform
