+++ examples/docs_snippets/docs_snippets/guides/components/index/2-e-pip-scaffold.txt
@@ -1,1 +1,1 @@
-dg init .
+dg scaffold project .
