+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/1-dg-scaffold-workspace.txt
@@ -1,0 +1,3 @@
+dg scaffold workspace dagster-workspace && cd dagster-workspace
+
+Scaffolded files for Dagster workspace at /.../dagster-workspace.
