+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/2-dg-scaffold-workspace.txt
@@ -1,0 +1,5 @@
+dg scaffold project --python-environment uv_managed projects/project-1
+
+Creating a Dagster project at /.../dagster-workspace/projects/project-1.
+Scaffolded files for Dagster project at /.../dagster-workspace/projects/project-1.
+...
