+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/3-tree.txt
@@ -1,0 +1,20 @@
+tree
+
+.
+├── dg.toml
+└── projects
+    └── project-1
+        ├── pyproject.toml
+        ├── src
+        │   └── project_1
+        │       ├── __init__.py
+        │       ├── definitions.py
+        │       ├── defs
+        │       │   └── __init__.py
+        │       └── lib
+        │           └── __init__.py
+        ├── tests
+        │   └── __init__.py
+        └── uv.lock
+
+...
