+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/4-dg.toml
@@ -1,0 +1,6 @@
+directory_type = "workspace"
+
+[workspace]
+use_editable_dagster = "dagster-workspace"
+[[workspace.projects]]
+path = "projects/project-1"
