+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/5-project-pyproject.toml
@@ -1,0 +1,7 @@
+...
+[tool.dg]
+directory_type = "project"
+
+[tool.dg.project]
+root_module = "project_1"
+...
