+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/6-scaffold-project.txt
@@ -1,0 +1,5 @@
+dg scaffold project projects/project-2 --python-environment uv_managed
+
+Creating a Dagster project at /.../dagster-workspace/projects/project-2.
+Scaffolded files for Dagster project at /.../dagster-workspace/projects/project-2.
+...
