+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/7-project-list.txt
@@ -1,0 +1,4 @@
+dg list project
+
+projects/project-1
+projects/project-2
