+++ examples/docs_snippets/docs_snippets/guides/dg/workspace/8-workspace.yaml
@@ -1,0 +1,9 @@
+load_from:
+  - python_file:
+      relative_path: projects/project-1/src/project_1/definitions.py
+      location_name: project_1
+      executable_path: projects/project-1/.venv/bin/python
+  - python_file:
+      relative_path: projects/project-2/src/project_2/definitions.py
+      location_name: project_2
+      executable_path: projects/project-2/.venv/bin/python
