+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/components/test_components_docs_creating_a_component.py
@@ -37,7 +37,7 @@ def test_creating_a_component(
     with isolated_snippet_generation_environment() as get_next_snip_number:
         # Scaffold code location
         _run_command(
-            cmd="dg init my-component-library --python-environment uv_managed --use-editable-dagster && cd my-component-library",
+            cmd="dg scaffold project my-component-library --python-environment uv_managed --use-editable-dagster && cd my-component-library",
         )
 
         #########################################################
