+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_list_commands.py
@@ -77,7 +77,7 @@ def test_list_projects_aliases(alias: str):
 def test_list_components_success():
     with (
         ProxyRunner.test(use_fixed_test_components=True) as runner,
-        isolated_example_project_foo_bar(runner, in_workspace=False),
+        isolated_example_project_foo_bar(runner, in_workspace=False, uv_sync=False),
     ):
         result = runner.invoke(
             "scaffold", "dagster_test.components.AllMetadataEmptyComponent", "qux"
@@ -574,7 +574,7 @@ def _sample_failed_defs():
 def test_list_env_succeeds():
     with (
         ProxyRunner.test(use_fixed_test_components=True) as runner,
-        isolated_example_project_foo_bar(runner, in_workspace=False),
+        isolated_example_project_foo_bar(runner, in_workspace=False, uv_sync=False),
     ):
         result = runner.invoke("list", "env")
         assert_runner_result(result)
