+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_scaffold_commands.py
@@ -1,10 +1,12 @@
 import json
+import os
 import shutil
 import subprocess
 import textwrap
 from pathlib import Path
-from typing import Any, Literal, get_args
+from typing import Any, Literal, Optional, get_args
 
+import dagster_shared.check as check
 import pytest
 import tomlkit
 from dagster_dg.cli.shared_options import DEFAULT_EDITABLE_DAGSTER_PROJECTS_ENV_VAR
@@ -45,30 +47,53 @@ from dagster_dg_tests.utils import (
 # ########################
 
 
-def test_scaffold_workspace_command_success(monkeypatch) -> None:
+@pytest.mark.parametrize(
+    "cli_args",
+    [
+        tuple(),
+        ("helloworld",),
+        (".",),
+    ],
+    ids=[
+        "no_args",
+        "with_name",
+        "with_cwd",
+    ],
+)
+def test_scaffold_workspace_command_success(monkeypatch, cli_args: tuple[str, ...]) -> None:
     with ProxyRunner.test() as runner, runner.isolated_filesystem():
-        result = runner.invoke("scaffold", "workspace")
+        if "." in cli_args:
+            os.mkdir("helloworld")
+            os.chdir("helloworld")
+            expected_name = "helloworld"
+        elif "helloworld" in cli_args:
+            expected_name = "helloworld"
+        else:
+            expected_name = "dagster-workspace"
+
+        result = runner.invoke("scaffold", "workspace", *cli_args)
         assert_runner_result(result)
-        assert Path("dagster-workspace").exists()
-        assert Path("dagster-workspace/dg.toml").exists()
-        assert Path("dagster-workspace/projects").exists()
-        assert not Path("dagster-workspace/libraries").exists()
 
-        result = runner.invoke("scaffold", "workspace")
-        assert_runner_result(result, exit_0=False)
-        assert "already exists" in result.output
+        if "." in cli_args:
+            os.chdir("..")
+
+        assert Path(expected_name).exists()
+        assert Path(f"{expected_name}/dg.toml").exists()
+        assert Path(f"{expected_name}/projects").exists()
 
 
-def test_scaffold_workspace_command_name_override_success(monkeypatch) -> None:
+def test_scaffold_workspace_already_exists_failure(monkeypatch) -> None:
+    dagster_git_repo_dir = discover_git_root(Path(__file__))
+    monkeypatch.setenv("DAGSTER_GIT_REPO_DIR", str(dagster_git_repo_dir))
+
     with ProxyRunner.test() as runner, runner.isolated_filesystem():
-        result = runner.invoke("scaffold", "workspace", "my-workspace")
-        assert_runner_result(result)
-        assert Path("my-workspace").exists()
-        assert Path("my-workspace/dg.toml").exists()
-        assert Path("my-workspace/projects").exists()
-        assert not Path("my-workspace/libraries").exists()
-
-        result = runner.invoke("scaffold", "workspace", "my-workspace")
+        os.mkdir("dagster-workspace")
+        result = runner.invoke(
+            "scaffold",
+            "workspace",
+            "--use-editable-dagster",
+            "dagster-workspace",
+        )
         assert_runner_result(result, exit_0=False)
         assert "already exists" in result.output
 
@@ -87,6 +112,81 @@ def test_scaffold_workspace_command_name_override_success(monkeypatch) -> None:
 # and returns the local version of the package.
 
 
+@pytest.mark.parametrize(
+    "cli_args,input_str,opts",
+    [
+        (("--", "."), "y\n", {}),
+        # Test preexisting venv in the project directory
+        (("--", "."), None, {"use_preexisting_venv": True}),
+        # Skip the uv sync prompt and automatically uv sync
+        (("--uv-sync", "--", "."), None, {}),
+        # Skip the uv sync prompt and don't uv sync
+        (("--no-uv-sync", "--", "."), None, {}),
+        # Test uv not available. When uv is not available there will be no prompt-- so this test
+        # will hang if it's not working because we don't provide an input string.
+        (("--", "."), None, {"no_uv": True}),
+        (("--", "foo-bar"), "y\n", {}),
+        # Test declining to create a venv
+        (("--", "foo-bar"), "n\n", {}),
+    ],
+    ids=[
+        "dirname_cwd",
+        "dirname_cwd_preexisting_venv",
+        "dirname_cwd_explicit_uv_sync",
+        "dirname_cwd_explicit_no_uv_sync",
+        "dirname_cwd_no_uv",
+        "dirname_arg",
+        "dirname_arg_no_venv",
+    ],
+)
+# def test_scaffold_project_outside_workspace_success(monkeypatch) -> None:
+def test_scaffold_project_success(
+    monkeypatch, cli_args: tuple[str, ...], input_str: Optional[str], opts: dict[str, object]
+) -> None:
+    use_preexisting_venv = check.opt_bool_elem(opts, "use_preexisting_venv") or False
+    no_uv = check.opt_bool_elem(opts, "no_uv") or False
+    # Remove when we are able to test without editable install
+    dagster_git_repo_dir = discover_git_root(Path(__file__))
+    monkeypatch.setenv("DAGSTER_GIT_REPO_DIR", str(dagster_git_repo_dir))
+    if no_uv:
+        monkeypatch.setattr("dagster_dg.cli.scaffold.is_uv_installed", lambda: False)
+    with ProxyRunner.test() as runner, runner.isolated_filesystem(), clear_module_from_cache("bar"):
+        if "." in cli_args:  # creating in CWD
+            os.mkdir("foo-bar")
+            os.chdir("foo-bar")
+            if use_preexisting_venv:
+                subprocess.run(["uv", "venv"], check=True)
+
+        # result = runner.invoke("scaffold", "project", "foo-bar", "--use-editable-dagster")
+        result = runner.invoke(
+            "scaffold", "project", "--use-editable-dagster", *cli_args, input=input_str
+        )
+        assert_runner_result(result)
+
+        if "." in cli_args:  # creating in CWD
+            os.chdir("..")
+
+        assert Path("foo-bar").exists()
+        assert Path("foo-bar/src/foo_bar").exists()
+        assert Path("foo-bar/src/foo_bar/lib").exists()
+        assert Path("foo-bar/src/foo_bar/defs").exists()
+        assert Path("foo-bar/tests").exists()
+        assert Path("foo-bar/pyproject.toml").exists()
+
+        # this indicates user opts to create venv and uv.lock
+        if not use_preexisting_venv and (
+            (input_str and input_str.endswith("y\n")) or "--uv-sync" in cli_args
+        ):
+            assert Path("foo-bar/.venv").exists()
+            assert Path("foo-bar/uv.lock").exists()
+        elif use_preexisting_venv:
+            assert Path("foo-bar/.venv").exists()
+            assert not Path("foo-bar/uv.lock").exists()
+        else:
+            assert not Path("foo-bar/.venv").exists()
+            assert not Path("foo-bar/uv.lock").exists()
+
+
 def test_scaffold_project_inside_workspace_success(monkeypatch) -> None:
     # Remove when we are able to test without editable install
     dagster_git_repo_dir = discover_git_root(Path(__file__))
@@ -98,6 +198,7 @@ def test_scaffold_project_inside_workspace_success(monkeypatch) -> None:
             "project",
             "projects/foo-bar",
             "--verbose",
+            "--uv-sync",
         )
         assert_runner_result(result)
         assert Path("projects/foo-bar").exists()
@@ -117,8 +218,8 @@ def test_scaffold_project_inside_workspace_success(monkeypatch) -> None:
         assert get_toml_node(toml, ("workspace", "projects", 0, "path"), str) == "projects/foo-bar"
 
         # Check venv not created
-        assert not Path("projects/foo-bar/.venv").exists()
-        assert not Path("projects/foo-bar/uv.lock").exists()
+        assert Path("projects/foo-bar/.venv").exists()
+        assert Path("projects/foo-bar/uv.lock").exists()
 
         # Restore when we are able to test without editable install
         # with open("projects/bar/pyproject.toml") as f:
@@ -175,26 +276,6 @@ def test_scaffold_project_inside_workspace_applies_scaffold_project_options(monk
         # Check that use_editable_dagster was applied
         toml = tomlkit.parse(Path("projects/foo-bar/pyproject.toml").read_text())
         assert has_toml_node(toml, ("tool", "uv", "sources", "dagster"))
-
-
-def test_scaffold_project_outside_workspace_success(monkeypatch) -> None:
-    # Remove when we are able to test without editable install
-    dagster_git_repo_dir = discover_git_root(Path(__file__))
-    monkeypatch.setenv("DAGSTER_GIT_REPO_DIR", str(dagster_git_repo_dir))
-
-    with ProxyRunner.test() as runner, runner.isolated_filesystem(), clear_module_from_cache("bar"):
-        result = runner.invoke("scaffold", "project", "foo-bar", "--use-editable-dagster")
-        assert_runner_result(result)
-        assert Path("foo-bar").exists()
-        assert Path("foo-bar/src/foo_bar").exists()
-        assert Path("foo-bar/src/foo_bar/lib").exists()
-        assert Path("foo-bar/src/foo_bar/defs").exists()
-        assert Path("foo-bar/tests").exists()
-        assert Path("foo-bar/pyproject.toml").exists()
-
-        # Check venv not created
-        assert not Path("foo-bar/.venv").exists()
-        assert not Path("foo-bar/uv.lock").exists()
 
 
 EditableOption: TypeAlias = Literal["--use-editable-dagster"]
@@ -267,7 +348,7 @@ def test_scaffold_project_use_editable_dagster_env_var_succeeds(monkeypatch) -> 
     with ProxyRunner.test() as runner, runner.isolated_filesystem():
         # We need to use subprocess rather than runner here because the environment variable affects
         # CLI defaults set at process startup.
-        subprocess.check_output(["dg", "scaffold", "project", "foo-bar"], text=True)
+        subprocess.check_output(["dg", "scaffold", "project", "--uv-sync", "foo-bar"], text=True)
         with open("foo-bar/pyproject.toml") as f:
             toml = tomlkit.parse(f.read())
             validate_pyproject_toml_with_editable(
